fix(admin-rewards): allow decimals in "Max % of bill" config field
Admin needed to set a fractional cap (e.g. 2.5%) but the input
parseInt'd every keystroke, so anything after the decimal point was
lost. Three coordinated changes:

- Schema: RewardConfig.maxRedemptionPctOfBill Int -> Float. Existing
  Int values (default 20) cast cleanly to 20.0 — non-breaking. The
  redemption math at lib/rewards/redeem.ts (Math.floor(billPaise *
  pct / 100)) was already float-safe.
- Zod: drop .int() on the field; min(0).max(100) still applies.
- NumberField: opt-in allowDecimals prop. When true, the input uses
  step="any", parseFloat, and a local draft so typing "2." doesn't
  snap to "2" before the next digit arrives. Every other field on
  the panel keeps the existing integer-only behavior.

+++ actions/admin-rewards.ts
@@ -155,7 +155,10 @@ const configSchema = z.object({
   earnRateCafeBps: z.number().int().min(0).max(10000),
   pointValuePaise: z.number().int().min(1).max(100000),
   minPointsToRedeem: z.number().int().min(0).max(1_000_000),
-  maxRedemptionPctOfBill: z.number().int().min(0).max(100),
+  // Float so admins can dial in fractional caps (2.5%, 12.5%, etc.)
+  // without bumping the schema to bps everywhere. The redemption math
+  // is float-safe (final paise number is floored).
+  maxRedemptionPctOfBill: z.number().min(0).max(100),
   maxRedemptionPaisePerTxn: z.number().int().min(0).max(10_000_000),
   pointExpiryMonths: z.number().int().min(1).max(120),
   earnToRedeemMinHours: z.number().int().min(0).max(24 * 365),
